revert prediction
